+++ src/Resources/assets/js/contao-flatpickr-bundle-theme.js
@@ -1,0 +1,1 @@
+import '../scss/flatpickr-bundle.scss';